Question
Given two non-negative integers n1 and n2, where n1 <n2. The task is to find the total number of integers
in the range interval [n1, n2] [both inclusive] which have no repeated digits.
For e.g.
Suppose n1 = 11 and n2 = 15.
There is the number 11, which has repeated digits, but 12, 13, 14, and 15 have no repeated digits. So, the
output is 4.